moved form state and logic to form component
