+++ components/searchForm.js
@@ -1,6 +1,15 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
-export default function SearchForm ({ handleSubmit, formData }) {
+export default function SearchForm ({ handleSubmit, setSearchQuery }) {
+    const [formData, setFormData] = useState({});
+
+    useEffect(() => {
+        const savedFormData = JSON.parse(sessionStorage.getItem('formData'));
+        if (savedFormData) {
+            setFormData(savedFormData);
+        }
+    }, []);
+
     const inputs = [
         { placeholder: 'Style', id: 'style', name: 'style', ariaLabel: 'style' },
         { placeholder: 'Year', id: 'year', name: 'year', ariaLabel: 'year' },
@@ -9,9 +18,25 @@ export default function SearchForm ({ handleSubmit, formData }) {
         { placeholder: 'Artist', id: 'artist', name: 'artist', ariaLabel: 'artist' },
     ];
 
+    const handleFormSubmit = async (e) => {
+        e.preventDefault();
+        const formData = new FormData(e.target);
+        const data = Object.fromEntries(formData);
+        if (Object.values(data).every(x => x === '')) {
+            setError({ message: 'Please enter search parameters' });
+            setFormData(data);
+            return;
+        }
+        Object.keys(data).forEach((key) => data[key] === '' && delete data[key]);
+        setFormData(data);
+        setSearchQuery(data);
+        sessionStorage.setItem('formData', JSON.stringify(data));
+    };
+
+
     return (
         <div className="search-form p-6">
-            <form onSubmit={handleSubmit}>
+            <form onSubmit={handleFormSubmit}>
                 <div className="flex flex-wrap justify-center align-center flex-col gap-1">
                     {inputs.map(input => (
                         <input
